Extract /list_patterns, /build pending, /build view, /search, /search_restrictions into new SearchCog
